+++ pyat/at/lattice/__init__.py
@@ -14,6 +14,7 @@ from .elements import *
 from .rectangular_bend import *
 from .idtable_element import InsertionDeviceKickMap
 from .utils import *
+from .transformation import *
 from .lattice_object import *
 # from .lattice_variables import *
 from .cavity_access import *
